\name{abundant-package}
\alias{abundant-package}
\alias{abundant}
\docType{package}
\title{
Abundant regression and high-dimensional principal fitted components
}
\description{
Fit and predict with the high-dimensional principal fitted components model.
}
\details{
The main functions are \code{\link{fit.pfc}}, \code{\link{pred.response}}.
}
\author{
Adam J. Rothman

Maintainer: Adam J. Rothman <arothman@umn.edu>
}
\references{
Cook, R. D., Forzani, L., and Rothman, A. J. (2012).  
Estimating sufficient reductions of the predictors in abundant high-dimensional regressions. 
Annals of Statistics 40(1), 353-384.
}


